pyproject.toml
[project]
name = "test-claude-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "google-auth>=2.49.1",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["test"]
addopts = "-v"
